+++ src/OpenSearch/HttpClient/HttpClientFactoryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenSearch;
+namespace OpenSearch\HttpClient;
 
 use Psr\Http\Client\ClientInterface;
 
@@ -16,6 +16,6 @@ interface HttpClientFactoryInterface
      *
      * @param array<string,mixed> $options
      */
-    public static function create(array $options): ClientInterface;
+    public function create(array $options): ClientInterface;
 
 }