+++ tests/HttpClient/SymfonyHttpClientFactoryTest.php
@@ -2,25 +2,25 @@
 
 declare(strict_types=1);
 
-namespace OpenSearch\Tests;
+namespace OpenSearch\Tests\HttpClient;
 
-use OpenSearch\SymfonyHttpClientFactory;
+use OpenSearch\HttpClient\SymfonyHttpClientFactory;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\ClientInterface;
 
 /**
  * Test the Symfony HTTP client factory.
  *
- * @coversDefaultClass \OpenSearch\SymfonyHttpClientFactory
+ * @coversDefaultClass \OpenSearch\HttpClient\SymfonyHttpClientFactory
  */
 class SymfonyHttpClientFactoryTest extends TestCase
 {
     public function testCreate()
     {
-        $client = SymfonyHttpClientFactory::create([
+        $factory = new SymfonyHttpClientFactory(2);
+        $client = $factory->create([
             'base_uri' => 'http://example.com',
             'verify_peer' => false,
-            'max_retries' => 2,
             'auth_basic' => ['username', 'password'],
         ]);